Prototype Android application
This prototype Android application is based on the new repository
hosting Measurement Kit JNI libs available at:

    https://github.com/bassosimone/measurement-kit-jni-libs

The implementation uses the JNI libs and their interface to run
in synchronous mode OONI tests. The basic mechanism to run in
background (inside a so-called service) and to notify back the
main thread when work is done is already implemented.