MainActivity.java: more robust resource copying (#79)
* NetworkMeasurement.java: initialize all attributes

Make sure all attributes are initialized, so to ensure that there
are no surprises if the constructor is not called.

It might be that the diff would be correct in C++ and correct and
overkill in Java, but what mattes is that it fixes #76.

Now, there is another pending diff to make sure we have completely
addressed all the issues exposed during #76.

* MainActivity.java: more robust resource copying

Rather than relying on a shared preference, directly check whether
the file is there or not and, in the latter case, copy it.

Observed during #76. Depends on #78.