Route notifications to a modal window in main activity
This diff adds useful abstraction to more easily route events through
the local broadcast manager, based on a previous diff by @AntonioLangiu
that I remember we were discussing on a taxi in Berlin.

This piece of work should actually be moved into MK's libraries
because it is generally useful. Also, the tests in this app should
probably be refactored to use this code.

Apart from that, I've modified the NotificationsRouter to send messages
through the local broadcast manager and the MainActivity to receive
such messages in its context and show the modal window.

From the graphic point of view, this could be, of course, improved a
lot. It may be possible, for example, to have a sort of "debug" activity
where we can read received messages.

Cc: @hellais.