CI: add support for github actions
When working on https://github.com/ooni/probe-engine/issues/893, I was
proceeding slower than I could because I had not enough confidence that
I was not breaking stuff. So, I'd say it's time to enable some form of
CI for this project. I also see that there is already a Travis file but
I also see that Travis is not running. Since we are already using GHA
for a bunch of client side projects, this strikes me as a good time to
switch this repository to using GHA.

Hopefully everything will work out of the box. If not, then we will
need to spend more time debugging after this first commit.

When everything is WAI, I'll purge .travis.yml.