CI: add support for github actions (#364)
When working on https://github.com/ooni/probe-engine/issues/893, I was
proceeding slower than I could because I had not enough confidence that
I was not breaking stuff. So, I'd say it's time to enable some form of
CI for this project. I also see that there is already a Travis file but
I also see that Travis is not running. Since we are already using GHA
for a bunch of client side projects, this strikes me as a good time to
switch this repository to using GHA.

Of course, let us also purge the .travis.yml file.