fix: use oonimkall 2021.01.29-110831
This corresponds to ooni/probe-engine v0.23.0-alpha.2-55-g23a9659.

In my local testing this commit is such that OONI Probe Dev does
not crash anymore as described in https://github.com/ooni/probe/issues/1303.

Because this fixes an issue, I'm marking this commit as a
fix rather than just as a chore.