feat: support tor and torsf proxies
Unfortunately, because of https://github.com/ooni/probe/issues/2406, we
are going to see crashes when using these proxies.

This diff is part of https://github.com/ooni/probe/issues/2500. Since we're
increasingly being blocked, it makes sense to exposes all the possible
proxies we can feature.

We're going to touch upon the same files again once we land the
https://github.com/ooni/probe-cli/pull/1162 pull request.